restructured program into multiple files for ease of use
